pyproject.toml
[project]
name = "imagine-penguin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "numpy>=2.3.5",
    "openai>=2.11.0",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pyinstaller>=6.17.0",
    "requests>=2.32.5",
]
